+++ frontend/src/reception/components/WaitingStatus.vue
