--- frontend/src/views/reception/Reception.vue
@@ -1,3 +1,0 @@
-<template>
-  <router-view />
-</template>
